requirements.txt
selenium~=4.15.2
PyYAML~=5.3.1
openpyxl~=3.1.2
pytest~=7.4.2